微调

+++ src/pageDetail/goodsManage/goodsManage.vue
@@ -86,25 +86,26 @@ export default {
     this.getsupplierList()
   },
   methods: {
-      getsupplierList(){
-      axios.post(supplierSelectAll).then(data=>{
-        console.log(data,'厂商')
-        this.supplierList=data
-      })
-    },
-    getcategoryList(){
-      axios.get(categoryList).then(data=>{
-        console.log(data);
-        this.categoryList=data;
-      })
-    },
+
     deleteBtn(row){
-      axios.delete(goodsDelete+'?id='+row).then(data=>{
-        this.$message.warning('删除成功')
-        this.getList();
-      })
-
-
+      this.$confirm('此操作将永久删除该商品, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete(goodsDelete+'?id='+row).then(data=>{
+          this.getList();
+        })
+        this.$message({
+          type: 'success',
+          message: '删除成功!'
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
+      });
 
     },
         //分页
